# Significant speedup of ASOF join