--- a/dbms/src/Interpreters/RowRefs.h
+++ b/dbms/src/Interpreters/RowRefs.h
@@ -1,12 +1,12 @@
 #pragma once
 
 #include <Columns/IColumn.h>
-#include <Common/SortedLookupPODArray.h>
 
 #include <optional>
 #include <variant>
 #include <list>
 #include <mutex>
+#include <algorithm>
 
 namespace DB
 {
@@ -32,14 +32,68 @@ struct RowRefList : RowRef
     RowRefList(const Block * block_, size_t row_num_) : RowRef(block_, row_num_) {}
 };
 
+/**
+ * This class is intended to push sortable data into.
+ * When looking up values the container ensures that it is sorted for log(N) lookup
+ * After calling any of the lookup methods, it is no longer allowed to insert more data as this would invalidate the
+ * references that can be returned by the lookup methods
+ */
+
+template <typename T>
+class SortedLookupVector
+{
+public:
+    using Base = std::vector<T>;
+
+    // First stage, insertions into the vector
+    template <typename U, typename ... TAllocatorParams>
+    void insert(U && x, TAllocatorParams &&... allocator_params)
+    {
+        assert(!sorted.load(std::memory_order_acquire));
+        array.push_back(std::forward<U>(x), std::forward<TAllocatorParams>(allocator_params)...);
+    }
+
+    // Transition into second stage, ensures that the vector is sorted
+    typename Base::const_iterator upper_bound(const T & k)
+    {
+        sort();
+        return std::upper_bound(array.cbegin(), array.cend(), k);
+    }
+
+    // After ensuring that the vector is sorted by calling a lookup these are safe to call
+    typename Base::const_iterator cbegin() const { return array.cbegin(); }
+    typename Base::const_iterator cend() const { return array.cend(); }
+
+private:
+    Base array;
+    std::atomic<bool> sorted = false;
+    mutable std::mutex lock;
+
+    // Double checked locking with SC atomics works in C++
+    // https://preshing.com/20130930/double-checked-locking-is-fixed-in-cpp11/
+    // The first thread that calls one of the lookup methods sorts the data
+    // After calling the first lookup method it is no longer allowed to insert any data
+    // the array becomes immutable
+    void sort()
+    {
+        if(!sorted.load(std::memory_order_acquire)) {
+            std::lock_guard<std::mutex> l(lock);
+            if(!sorted.load(std::memory_order_relaxed)) {
+                std::sort(array.begin(), array.end());
+                sorted.store(true, std::memory_order_release);
+            }
+        }
+    }
+};
+
 class AsofRowRefs
 {
 public:
     template <typename T>
     struct Entry
     {
-        using LookupType = SortedLookupPODArray<Entry<T>>;
-
+        using LookupType = SortedLookupVector<Entry<T>>;
+        using LookupPtr = std::unique_ptr<LookupType>;
         T asof_value;
         RowRef row_ref;
 
@@ -53,16 +107,10 @@ public:
     };
 
     using Lookups = std::variant<
-        Entry<UInt32>::LookupType,
-        Entry<UInt64>::LookupType,
-        Entry<Float32>::LookupType,
-        Entry<Float64>::LookupType>;
-
-    struct LookupLists
-    {
-        mutable std::mutex mutex;
-        std::list<Lookups> lookups;
-    };
+        Entry<UInt32>::LookupPtr,
+        Entry<UInt64>::LookupPtr,
+        Entry<Float32>::LookupPtr,
+        Entry<Float64>::LookupPtr>;
 
     enum class Type
     {
@@ -72,13 +120,23 @@ public:
         keyf64,
     };
 
+    AsofRowRefs() {}
+    AsofRowRefs(Type t);
+
     static std::optional<Type> getTypeSize(const IColumn * asof_column, size_t & type_size);
 
-    void insert(Type type, LookupLists &, const IColumn * asof_column, const Block * block, size_t row_num);
-    const RowRef * findAsof(Type type, const LookupLists &, const IColumn * asof_column, size_t row_num) const;
+    // This will be synchronized by the rwlock mutex in Join.h
+    void insert(Type type, const IColumn * asof_column, const Block * block, size_t row_num);
+
+    // This will internally synchronize
+    const RowRef * findAsof(Type type, const IColumn * asof_column, size_t row_num) const;
 
 private:
-    Lookups * lookups = nullptr;
+    // Lookups can be stored in a HashTable because it is memmovable
+    // A std::variant contains a currently active type id (memmovable), together with a union of the types
+    // The types are all std::unique_ptr, which contains a single pointer, which is memmovable.
+    // Source: https://github.com/yandex/ClickHouse/issues/4906
+    Lookups lookups;
 };
 
 }