I hereby agree to the terms of the CLA available at: https://yandex.ru/legal/cla/?lang=en

For changelog. 
Speed up multicore performance of ASOF join by removing the lock per join and replacing it with double checked locking per join key.

Category (leave one):
- Performance Improvement

**Short description (up to few sentences):**
Previously all of the AsofRowRefs were protected by a single mutex per Join object, which is obviously not ideal from a performance perspective as joinBlock becomes essentially single threaded (for the non-asof case this can be called by multiple threads simultaneously on their own blocks as they do not modify the state of the Join object). I moved the lock into the SortedLookupVector instead, which distributes them between the different keys. The reason for the need of the lock is that the AsofRowRefs only sorts it's contents when it's required (and therefore isn't strictly read-only, but is accessed in the JoinBlock stage of the Join). However, it is logically read-only, as the user should not observe any changes, merely the faster lookup.

**Code locations to be aware of**
1. Double-checked locking
The global lock is removed in favour of a lock per SortedLookupVector. This used to be in Common/SortedLookupPODArray.h, though the locking strategy is very specialized to be used in the Join.h, therefore I've moved it more locally. As an example, it does not lock upon insert (as it relies on the shared_mutex in Join.h to synchronize that) and also it is undefined behaviour to insert after the container has been sorted, further adding caveats and reducing generality. The check whether to actually sort the data is using double-checked locking, based on (https://preshing.com/20130930/double-checked-locking-is-fixed-in-cpp11/#using-c11-low-level-ordering-constraints), so on the fast path it reduces to an uncontested atomic load. 

2. Storing the AsofRowRefs object directly in the HashTable.
Similar to std::vector, AsofRowRefs is now memmovable (as per https://github.com/yandex/ClickHouse/issues/4906). Therefore it's stored directly in the HashTable rather than storing a pointer to it and the object itself living in a per-join std::list containing all of the AsofRowRefs.

**Quick benchmark**
A quick benchmark (ran on my laptop, but the difference is significant enough that it should be sufficient evidence) shows a ~4x improvement. This is expected to significantly increase on servers with more cores available than I have for testing (from `00927_asof_join_big.sql`, which performs asof join on 2 time series containing resp. 15k and 50k data points per key, with 5k keys). This should be a reasonable synthetic performance benchmark for now:

``` 
SELECT SUM(trades.price - tvs.tv) FROM trades ASOF LEFT JOIN tvs USING(k,t);
```

```
martijn-asof-speedup (double checked locking per AsofRowRef)
1 rows in set. Elapsed: 10.426 sec. Processed 325.00 million rows, 5.20 GB (31.17 million rows/s., 498.76 MB/s.)

master (add all to a single std::list and a single mutex)
1 rows in set. Elapsed: 44.752 sec. Processed 325.00 million rows, 5.20 GB (7.26 million rows/s., 116.20 MB/s.)
```