@Gladdy 2019-04-06
I'm a bit puzzled why the latter two builds fail some of the performance tests. A quick inspection shows that moving the atomic has some performance impact even when the test doesn't even perform an asof join, which indicates that the failures have nothing to do with my changes and are just flukes?