@4ertus2 2019-04-06
> are just flukes?

Probably they are. Auto perf tests are new and not stable enough.