@4ertus2 2019-04-08
I've changed and renamed 00927_asof_join_long.sql
It triggered tests timeouts and out of memory/disk for test environment.
Names of big tests should contain word 'long' cause there's option --no-long for ctest run.