@Gladdy 2019-04-08
Thanks, sorry, I missed that. Might also be wise to just move it with the other performance tests which I didn't get around to yet.